requirements.txt
numpy==2.2.3
onnxruntime==1.20.1
Pillow==11.1.0
torch==2.6.0+cu126
torchvision==0.21.0
